Fix issue with AsciiPageReader
It would fail if the input file was opened in 'rb' mode, which was the
default mode to open files fgi-osnma.

Also, PageReader.read will now return just the INAV page, and no longer
the system timestamp of the page arrival. The timestamp of arrival was
moved inside the InavPage object.